+++ react/pages/_app.js
@@ -1,5 +1,5 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
-import '@/styles/newglobals.css'
+import '@/styles/globals.css'
 import '@/styles/404.css'
 
 export default function App({ Component, pageProps }) {
